[#33] Work-around for bug in Maven
When including cloudkeeper-parent as a submodule in another project, it
was observed that the integration test for cloudkeeper-maven-plugin may
fail with:

java.lang.IncompatibleClassChangeError:
org/eclipse/sisu/plexus/scanners/PlexusTypeVisitor

The problem was that the JAR for org.ow2.asm:asm could appear on the
classpath that the integration test for cloudkeeper-maven-plugin was run
with. Presumably, the Maven reactor under some circumstances sees the
pre-Shade POM of cloudkeeper-asm-repackaged-5.0. Oddly though, this
problem does not occur when just building cloudkeeper-parent.

Steps to reproduce:
- Run ‘mvn clean’ on the parent project that contains
  project cloudkeeper-parent as a submodule.
- Remove all CloudKeeper artifacts from the Maven local repository, and
  run ‘mvn -B install -X -e’.
- Under directory
  cloudkeeper/cloudkeeper-maven/cloudkeeper-maven-plugin/target/surefire
  there should be two files ending in ‘tmp’, and one contains a list of
  classpaths. Prior to this fix, asm-5.0.4.jar was (sometimes) contained
  in the classpath, and the test would fail (as legacy Maven code
  requires version 3 of asm).