[#34] Update AWS SDK to 1.10.77
Updated AWS SDK from 1.8.9 to 1.10.77 and replaced dependency on
aws-java-sdk with the smaller aws-java-sdk-s3. Also updated dependency
on Maven to 3.3.9. Eclipse Aether needs to stay at version 1.0.2 for
now, because the only later version 1.1.0 changes dependency resolution
in a way that should be considered a bug.

Version management of httpclient is now necessary because both AWS and
Aether bring it as transitive dependency.